refactor: split user settings api clients
